Replace interactive layout with 3D and remove static script
